move distance generation logic into separate files
